cgroup2 support
Refactored how the Dockside container ID is identified. The method for
cgroup v1 is not applicable for cgroup v2, unless the docker run
--cgroupns=host option is used.

Without that option, alternative methods now include:
- docker run --cidfile option passing container id to bind-mounted fs
- obtaining the container's hostname from $HOSTNAME

N.B. The Dockside container ID is required to identify the IDE volume
when bind-mounting /var/run/docker.dock, but is not required when
running an inner dockerd, e.g. with the Sysbox runtime.